reorg of overlay

+++ ris_widget/overlay/point_set.py
@@ -8,13 +8,16 @@ from . import base
 class _PointHandle(base.SelectableHandle):
     QGRAPHICSITEM_TYPE = shared_resources.generate_unique_qgraphicsitem_type()
 
-    def __init__(self, parent, layer_stack, color):
-        super().__init__(parent, layer_stack, color)
+    def __init__(self, point_set, pos, brush, pen=None):
+        self._geometry_changed = point_set._geometry_changed
+        layer_stack = point_set.parentItem()
+        super().__init__(layer_stack, layer_stack, brush, pen)
+        self.setPos(pos)
         self.setFlag(Qt.QGraphicsItem.ItemSendsGeometryChanges) # Necessary in order for .itemChange to be called when item is moved
 
     def itemChange(self, change, value):
         if change == Qt.QGraphicsItem.ItemPositionHasChanged:
-            self.parentItem()._geometry_changed()
+            self._geometry_changed()
         else:
             return super().itemChange(change, value)
 
@@ -23,11 +26,15 @@ class PointSet(base.RWGeometryItemMixin, Qt.QGraphicsPathItem):
     POINT_TYPE = _PointHandle
 
     # just need to inherit from some QGraphicsItem that can have a pen set
-    def __init__(self, ris_widget, pen=None, geometry=None, max_points=None):
+    def __init__(self, ris_widget, brush=None, pen=None, geometry=None, max_points=None):
         self.max_points = max_points
         self.points = []
         self._last_click_deselected = False
-        super().__init__(ris_widget, pen, geometry)
+        if brush is None:
+            brush = Qt.Qt.green
+        self.brush = brush
+        self.pen = pen
+        super().__init__(ris_widget, geometry=geometry)
 
     @property
     def geometry(self):
@@ -47,26 +54,26 @@ class PointSet(base.RWGeometryItemMixin, Qt.QGraphicsPathItem):
         self.points = []
         if geometry is not None:
             for x, y in geometry:
-                self._add_point(Qt.QPointF(x, y), skip_change=True)
-        self._geometry_changed()
+                self._add_point(Qt.QPointF(x, y))
 
-    def _add_point(self, pos, skip_change=False):
+    def _add_point(self, pos):
         if self.max_points is None or len(self.points) < self.max_points:
-            point = self.POINT_TYPE(self, self.parentItem(), self.pen().color())
-            point.setPos(pos)
-            self.points.append(point)
-            if not skip_change:
-                self._geometry_changed()
+            self.points.append(self.POINT_TYPE(self, pos, self.brush, self.pen))
+            return True
+        return False
 
     def _delete_selected(self):
         new_points = []
+        deleted = False
         for point in self.points:
             if point.isSelected():
                 point.remove()
+                deleted = True
             else:
                 new_points.append(point)
         self.points = new_points
-        self._geometry_changed()
+        if deleted:
+            self._geometry_changed()
 
     def remove(self):
         for point in self.points:
@@ -74,10 +81,11 @@ class PointSet(base.RWGeometryItemMixin, Qt.QGraphicsPathItem):
         super().remove()
 
     def _view_mouse_release(self, pos, modifiers):
-        # Called when ROI item is visible, and a mouse-up on the underlying
+        # Called when item is visible, and a mouse-up on the underlying
         # view occurs. (I.e. not on this item itself)
         if not self._last_click_deselected:
-            self._add_point(pos)
+            if self._add_point(pos):
+                self._geometry_changed()
 
     def sceneEventFilter(self, watched, event):
         event_type = event.type()
@@ -87,7 +95,8 @@ class PointSet(base.RWGeometryItemMixin, Qt.QGraphicsPathItem):
             else:
                 self._last_click_deselected = False
             # don't return true to not swallow the mouse click
-        elif event_type == Qt.QEvent.KeyPress and event.key() in {Qt.Qt.Key_Delete, Qt.Qt.Key_Backspace}:
+        elif (event_type == Qt.QEvent.KeyPress and event.key() in {Qt.Qt.Key_Delete, Qt.Qt.Key_Backspace} and
+                any(point.isSelected() for point in self.points)):
             self._delete_selected()
             return True
         return False
